--- token/services/identity/driver/wallet.go
@@ -1,7 +1,0 @@
-/*
-Copyright IBM Corp. All Rights Reserved.
-
-SPDX-License-Identifier: Apache-2.0
-*/
-
-package driver
