+++ token/services/identity/role/role.go
@@ -4,7 +4,7 @@ Copyright IBM Corp. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package common
+package role
 
 import (
 	"runtime/debug"
